Object loading for the Editor. object file is now included in the level archive. Massive GUI-update. AWT-File Dialogs instead of swing now being used. Magic button broken. Toolbar with cool buttons. Transparent Grid. Basic implementation of a selection tool. Camera movement through selection mode. Little tweaks on Mario. Camera position can now be forced to a specific position.

+++ trunk/2d-java-game-engine/src/mapeditor/GameObject.java
@@ -62,7 +62,7 @@ public class GameObject {
 
     @Override
     public String toString(){
-        return name + " = " + objectChar;
+        return name + " = " + (int)objectChar;
     }
 
 }
